+++ src/services/PaymentSources.php
@@ -127,6 +127,31 @@ class PaymentSources extends Component
     }
 
     /**
+     * Returns all payment sources for a gateway.
+     *
+     * @param int|null $gatewayId the gateway's ID
+     * @return PaymentSource[]
+     */
+    public function getAllPaymentSourcesByGatewayId(int $gatewayId = null): array
+    {
+        if (null === $gatewayId) {
+            return [];
+        }
+
+        $results = $this->_createPaymentSourcesQuery()
+            ->where(['gatewayId' => $gatewayId])
+            ->all();
+
+        $sources = [];
+
+        foreach ($results as $result) {
+            $sources[] = new PaymentSource($result);
+        }
+
+        return $sources;
+    }
+
+    /**
      * Returns a user's payment sources on a gateway, per the user's ID.
      *
      * @param int|null $gatewayId the gateway's ID