+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added a “Variants” column to the Product index page. ([#1765](https://github.com/craftcms/commerce/issues/1765))
+- Added `craft\commerce\services\PaymentSources::getAllPaymentSourcesByGatewayId()`.
 
 ### Changed
 - Disabled the browser autosuggest list when searching for a customer on the Edit Order page. ([#1752](https://github.com/craftcms/commerce/issues/1752))
@@ -13,6 +14,7 @@
 - Fixed a bug where the store location address had it‘s `isStoreLocation` property set to `false`. ([#1773](https://github.com/craftcms/commerce/issues/1773))
 - Fixed a bug where template suggestions weren’t showing on the product type template field.
 - Fixed a bug where deleting products would not delete related disabled variant. ([#1772](https://github.com/craftcms/commerce/issues/1772))
+- Fixed a bug during payment where carts could use an archived gateway. ([#1531](https://github.com/craftcms/commerce/issues/1531)) 
 
 ## 3.2.7 - 2020-09-24